Rename and clean up files.
- Rename all header files with "tapp_" prefix. This will help prevent problems when users do something like `#include "tapp_XXX.h"` with `-I/path/to/tapp/tapp`.
- Remove some old discussion and proposal files.
