+++ src/tapp.h
@@ -1,13 +1,13 @@
 #ifndef TAPP_TAPP_H_
 #define TAPP_TAPP_H_
 
-#include "tapp/error.h"
-#include "tapp/attributes.h"
-#include "tapp/datatype.h"
-#include "tapp/handle.h"
-#include "tapp/executor.h"
-#include "tapp/status.h"
-#include "tapp/tensor.h"
-#include "tapp/product.h"
+#include "tapp/tapp_error.h"
+#include "tapp/tapp_attributes.h"
+#include "tapp/tapp_datatype.h"
+#include "tapp/tapp_handle.h"
+#include "tapp/tapp_executor.h"
+#include "tapp/tapp_status.h"
+#include "tapp/tapp_tensor.h"
+#include "tapp/tapp_product.h"
 
 #endif /* TAPP_TAPP_H_ */
